random tables page working
All random table page features are now working.

Still need to make sure that the data structure outputs all of it's data to the form for character and calendar pages when an xml file is loaded. (It deSerializes from the file into the data structure, but no idea how/if those pages are triggering a load to the form). Ditto for saving to the data structure from the form for calendar and character pages.